lhdcv5: Change the type of bitrate tables to &[u32]
Previous type was i32, all bitrate values are positive
and using i32 does not make sense

Bug: 454096420
Test: m com.android.bt
Flag: EXEMPT PURE_REFACTOR

+++ system/audio/codecs/lhdcv5/src/lhdc_api.rs
@@ -135,14 +135,13 @@ fn MIN_BITRATE_LIMIT(X: i32, Y: i32) -> i32 {
     }
 }
 
-pub static g_bitrate_table_44k: [i32; 15] =
+pub static g_bitrate_table_44k: [u32; 15] =
     [64, 160, 192, 240, 320, 400, 480, 900, 1000, 1100, 1200, 1300, 1400, 99999, 1536000];
-
-pub static g_bitrate_table_48k: [i32; 15] =
+pub static g_bitrate_table_48k: [u32; 15] =
     [64, 160, 192, 256, 320, 400, 500, 900, 1000, 1100, 1200, 1300, 1400, 99999, 1536000];
-pub static g_bitrate_table_96k: [i32; 15] =
+pub static g_bitrate_table_96k: [u32; 15] =
     [64, 160, 192, 256, 320, 400, 500, 900, 1000, 1100, 1200, 1300, 1400, 99999, 1536000];
-pub static g_bitrate_table_192k: [i32; 15] =
+pub static g_bitrate_table_192k: [u32; 15] =
     [64, 160, 192, 256, 320, 400, 500, 900, 1000, 1100, 1200, 1300, 1400, 99999, 1536000];
 
 //jimmy
@@ -391,7 +390,7 @@ pub fn lhdcv5_enc_util_enc_process(
 pub fn lhdcv5_enc_util_get_bitrate(
     bitrate_inx: u32,
     bitrate: &mut u32,
-    bitrate_table: &[i32],
+    bitrate_table: &[u32],
 ) -> i32 {
     if bitrate_inx as usize >= bitrate_table.len() {
         error!("Input bit rate (index) is out of range ({bitrate_inx})!");
@@ -408,9 +407,9 @@ pub fn lhdcv5_enc_util_get_bitrate(
 pub fn lhdcv5_enc_util_get_bitrate_inx(
     bitrate: u32,
     bitrate_inx: &mut u32,
-    bitrate_table: &[i32],
+    bitrate_table: &[u32],
 ) -> i32 {
-    if bitrate > bitrate_table[bitrate_table.len() - 1] as libc::c_uint {
+    if bitrate > bitrate_table[bitrate_table.len() - 1] {
         return LHDC_FRET_INVALID_INPUT_PARAM;
     }
     let func_ret = lhdcv5_encoder_get_bitrate_inx(bitrate, bitrate_inx, bitrate_table);
